% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{bedfromfile}
\alias{bedfromfile}
\title{Input the bed format data.frame and convert to grange}
\usage{
bedfromfile(data, name = "name", strand = NULL)
}
\arguments{
\item{data}{input the data.frame}

\item{name}{assign the fourth column name and keep it exsist in grange}

\item{strand}{if strand assign as TRUE means input data contain the strand
information at column 6, if assign NULL strand will be *, otherwise could be "+" or "-"}
}
\description{
Input the bed format data.frame and convert to grange
}
