% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{binomialPeakCompile}
\alias{binomialPeakCompile}
\title{Compare the observe compile information with expect compile information by binomial distribution}
\usage{
binomialPeakCompile(
  observe,
  expect.data = expect.data,
  parallel = 1,
  parallel.type = "mclapply"
)
}
\arguments{
\item{observe}{The observe compile information}

\item{expect.data}{The expect compile information}

\item{parallel}{The number of cores to run the function}

\item{parallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Use mclapply to run in parallel\cr
\cr
"bplapply" - Use BiocParallel to run in parallel}
}
\description{
Compare the observe compile information with expect compile information by binomial distribution
}
