% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{BinsDefine}
\alias{BinsDefine}
\title{Given the input of bin size, minimum, and maximum values, generate intervals from the minimum to the maximum in specified bin increments. 
This string could be used in counting the occurrences of factors with associated elements.}
\usage{
BinsDefine(bin = NULL, min = NULL, max = NULL, type = NULL)
}
\arguments{
\item{bin}{Interval size of each window..}

\item{min}{Minimum number of conditions}

\item{max}{Maximum number of conditions}

\item{type}{Could be specify: \cr
\cr
"continue" - Output a string of intervals with the range of distances between 0 and a series of consecutive numbers. \cr
\cr
"eachBin"   - Output a string of intervals based on specified bin sizes, dividing the range of distances into equal segments.}
}
\description{
Given the input of bin size, minimum, and maximum values, generate intervals from the minimum to the maximum in specified bin increments. 
This string could be used in counting the occurrences of factors with associated elements.
}
