% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromDist.R
\name{ChromSTAT}
\alias{ChromSTAT}
\title{Input the data processed to calculate the correlation between features and regions for each chromosome.}
\usage{
ChromSTAT(data, parallel = 1)
}
\arguments{
\item{data}{Input object}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}
}
\description{
Input the data processed to calculate the correlation between features and regions for each chromosome.
}
