% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromDist.R
\name{CombineChrom}
\alias{CombineChrom}
\title{Function for the combination of the results across all chromosome.}
\usage{
CombineChrom(data)
}
\arguments{
\item{data}{Input object}
}
\description{
Function for the combination of the results across all chromosome.
}
