% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchieveFunction.R
\name{compileCorrelationByBin}
\alias{compileCorrelationByBin}
\title{Input shuffle peaks correlated (annotation) information as a list object.}
\usage{
compileCorrelationByBin(
  dir,
  parallel = FALSE,
  shuffle.n = 10000,
  shuffle.prefix = "shuffle_",
  file.ext = ".txt.gz",
  intersect = TRUE,
  bin = 1000,
  min = 0,
  max = 1e+06,
  count.type = "continue"
)

compileCorrelationByBin(
  dir,
  parallel = FALSE,
  shuffle.n = 10000,
  shuffle.prefix = "shuffle_",
  file.ext = ".txt.gz",
  intersect = TRUE,
  bin = 1000,
  min = 0,
  max = 1e+06,
  count.type = "continue"
)
}
\arguments{
\item{dir}{Directory of shuffle peak sets}

\item{parallel}{If set TRUE, will compile the shuffle peak sets with multiple threads}

\item{shuffle.n}{Number of peaks input into list. Reflect the flag (1~n) of shuffle_$flag.txt.gz}

\item{shuffle.prefix}{Name of prefix of peak sets.}

\item{file.ext}{Extension name of peak sets.}

\item{intersect}{If assign TRUE, output result will contained the number intersect peaks ($tag_dist==0) at first row}

\item{condition}{list of two number saperate with "-", will be used to count the number of factor A within the distance of numbers.}
}
\description{
Input shuffle peaks correlated (annotation) information as a list object.

Input shuffle peaks correlated (annotation) information as a list object.
}
