% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{CompileInfo}
\alias{CompileInfo}
\title{Transform the data contained associated factors to each element with distance to a data list}
\usage{
CompileInfo(data, dist = 1e+06, intersect = FALSE, include = "all")
}
\arguments{
\item{data}{The data contained associated factors to each element with distance}

\item{dist}{The distance to define the associated factors}

\item{intersect}{If  TRUE, results will be the number of elements that intersect with the factor}

\item{include}{Could be specified one of the: \cr
\cr
"all" - Include all the elements that intersect with the factor. \cr
\cr
"upstream" - Include all the elements that at the upstream of the factor. \cr
\cr
"downstream" - Include all the elements that at the downstream of the factor.}
}
\description{
Transform the data contained associated factors to each element with distance to a data list
}
