% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{CompilePeak}
\alias{CompilePeak}
\title{Compile the distance information of factors to elements within distance}
\usage{
CompilePeak(
  query,
  subject = subject,
  name = "name",
  DistIn = 1e+06,
  parallel = FALSE,
  save = NULL,
  strand = NULL
)
}
\arguments{
\item{query}{input query peak}

\item{subject}{input subject peak}

\item{name}{assign the fourth column name and keep it exsist in grange}

\item{DistIn}{assign the distance threshold}

\item{parallel}{if assign TRUE, the function will run in parallel}

\item{save}{if assign TRUE, the function will save the result}

\item{strand}{if strand assign as TRUE means input data contain the strand information at column 6, if assign NULL strand will be *, otherwise could be "+" or "-"}
}
\description{
Compile the distance information of factors to elements within distance
}
