% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{compileStat}
\alias{compileStat}
\title{Apply the statistic model to observe and expect peaks}
\usage{
compileStat(
  observe = observe,
  expect = expect,
  stat.type = "both",
  tail = "both",
  plotINFO.save = FALSE,
  log.p = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{observe}{Information of observed peaks}

\item{expect}{Information of expect peak sets list.}

\item{stat.type}{Statistic method. \cr
norm: Applied p.norm to calculate the}
}
\description{
Use expected peak set result to build the empirical model and compare observe number to model to calculate the significance.
}
