% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{CountCorrelation}
\alias{CountCorrelation}
\title{Input the data processed by `RegionAnnoFromData`` and tally the occurrences for each condition 
(e.g., intersections or specific distance ranges).}
\usage{
CountCorrelation(
  data,
  intersect = TRUE,
  condition = c("0-3000", "3000-10000", "10000-20000", "20000-30000", "30000-40000",
    "40000-50000")
)
}
\arguments{
\item{data}{Input data.frame or compressed/uncompressed file. 
Data should be in the format outputted by RegionAnnoFromData . It contain the information of: \cr
\cr
column 1: Name of the interval for the input factor. \cr
column 2: Name of the interval for the input element. \cr
column 3: Relative position information of factors with respect to the nearest element. \cr
column 4: Distance from the factors to the nearest element.}

\item{intersect}{If set to TRUE, the output result will place the number of intersecting peaks 
(where column 4's "distance" is equal to 0) in the first row.}

\item{condition}{A list of two numbers separated by a hyphen ("-"). Input the annotated data and tally the 
occurrences for each condition (e.g., intersections or specific distance ranges).}
}
\description{
Input the data processed by `RegionAnnoFromData`` and tally the occurrences for each condition 
(e.g., intersections or specific distance ranges).
}
