% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse.R
\name{CountCorrelationByBin}
\alias{CountCorrelationByBin}
\title{Input the correlated (annotation) data}
\usage{
CountCorrelationByBin(
  data,
  intersect = TRUE,
  bin = 1000,
  min = 0,
  max = 1e+06,
  count.type = "continue"
)
}
\arguments{
\item{data}{Input data.frame or compressed/uncompressed data. Data should output from RegionAnnoFromData, contain the information of
name (1st column); $tag_region (2nd column); $tag_type (3rd column); $tag_dist (4th column)}

\item{intersect}{If assign TRUE, output result will contained the number intersect peaks ($tag_dist==0) at first row}

\item{bin}{Interval size of each window.}

\item{min}{Minimum number of conditions}

\item{max}{Maximum number of conditions}

\item{type}{Could be specify one of: \cr
\cr
"continue" - All conditions will continues from min+interval*(max/bin(n)-1) to min+interval(max/bin(n))\cr
\cr
"within"   - All conditions will start from 0 to each intervals {0+interval(max/bin(n))}}
}
\description{
Input the correlated (annotation) data and count the number of each condition by bins. (e.g. intersect or with specific distances)
}
