% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchieveCompilePeak.R
\name{DistCalculate}
\alias{DistCalculate}
\title{Extract distance between query and subject within a distance threshold}
\usage{
DistCalculate(
  query,
  subject = subject,
  name = "name",
  DistIn = 1e+06,
  OVERLAY.ONLY = FALSE,
  strand = NULL,
  verbose = FALSE
)
}
\arguments{
\item{query}{input query peak}

\item{subject}{input subject peak}

\item{name}{assign the fourth column name and keep it exsist in grange}

\item{DistIn}{assign the distance threshold}

\item{strand}{if strand assign as TRUE means input data contain the strand information at column 6, if assign NULL strand will be *, otherwise could be "+" or "-"}
}
\description{
Extract distance between query and subject within a distance threshold
}
