% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchieveFunction.R
\name{ExtractCorrelate}
\alias{ExtractCorrelate}
\title{Output the distance of factor X to nearest element A.}
\usage{
ExtractCorrelate(
  data,
  txdb = txdb,
  tag = NULL,
  name = "name",
  strand = NULL,
  verbose = TRUE,
  strand.INFO = FALSE
)
}
\arguments{
\item{data}{Input factor X as bed format file, data.frame or grange.}

\item{txdb}{Element A as txdb format used for X annotation}

\item{tag}{tag before the names of column of output data}

\item{name}{ID of factor X used for annotation}

\item{strand}{if strand assign as TRUE means input data contain the strand information at column 6, if assign NULL strand will be *, otherwise could be "+" or "-"}

\item{verbose}{if TRUE, output the detail of processing}

\item{strand.INFO}{if TRUE, means the name of txdb contained strand information ("/+" and "/-")}
}
\description{
Extract the correlated information from factor X to txdb object made by element A.
}
