% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchievePeakStat.R
\name{ExtractDistance}
\alias{ExtractDistance}
\title{Extract the distance information from a vector.}
\usage{
ExtractDistance(data, dist = dist, intersect = TRUE)
}
\arguments{
\item{data}{Input data.}

\item{dist}{Distance to calculate the number of peaks.}

\item{intersect}{If TRUE, will extract the distance smaller than dist. If FALSE, will extract the distance larger than 0 and smaller than dist.}
}
\description{
Extract the distance information from a vector.
}
