% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{ExtractStartEnd}
\alias{ExtractStartEnd}
\title{Extract start and end position from grange}
\usage{
ExtractStartEnd(grange, strand = "*")
}
\arguments{
\item{grange}{input grange}
}
\description{
Extract start and end position from grange
}
