% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{FactorElementCorObj}
\alias{FactorElementCorObj}
\title{Compile the associated factors to each element with distance information}
\usage{
FactorElementCorObj(
  element,
  factor = factor,
  dist = 1e+06,
  strand = FALSE,
  parallel = 1,
  parallel.type = "mclapply"
)
}
\arguments{
\item{element}{The element file with bed format}

\item{factor}{The factor file with bed format}

\item{dist}{The distance to the element for collection}

\item{strand}{If TRUE, modified the dist with strand information}
}
\description{
Compile the associated factors to each element with distance information
}
