% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{FactorElementCorrelate}
\alias{FactorElementCorrelate}
\title{Input factors and associate them with elements. The output data.frame contains information 
about factors and their nearest elements with the distance.}
\usage{
FactorElementCorrelate(
  factor,
  element = element,
  strand = FALSE,
  tag = FALSE,
  outloc = NULL
)
}
\arguments{
\item{factor}{Input the factor data.frame or the path to the bed file.}

\item{element}{Input the element data.frame or the path to bed file.}

\item{strand}{If set to TRUE, it means that the input element contains strand information 
in column 6, and the analysis will take the strand information into consideration.}

\item{tag}{If a character is assigned, the tag information will be outputted in the final table column.}

\item{outloc}{The location of the output file.}
}
\description{
Input factors and associate them with elements. The output data.frame contains information 
about factors and their nearest elements with the distance.
}
