% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{FactorShufCorrelate}
\alias{FactorShufCorrelate}
\title{The input factors are shuffled and associated with elements. 
The output data.frame contains information about the random shuffle factors and their nearest elements with the distance.}
\usage{
FactorShufCorrelate(
  factor,
  element = element,
  strand = FALSE,
  tag = FALSE,
  outloc = NULL,
  genome = genome,
  incl = NULL,
  excl = NULL,
  seed = 1
)
}
\arguments{
\item{factor}{Input the factor data.frame or the path to the bed file.}

\item{element}{Input the element data.frame or the path to bed file.}

\item{strand}{If set to TRUE, it means that the input element contains strand information in column 6, 
and the analysis will take the strand information into consideration.}

\item{tag}{If a character is assigned, the tag information will be outputted in the final table column.}

\item{outloc}{The location of the output file.}

\item{genome}{This parameter specifies the data or file path containing the names and sizes of chromosomes or contigs. 
Each name-size pair should be listed on a separate line and delimited by a tab.}

\item{incl}{The interval information to include the input regions.}

\item{excl}{The interval information to exclude the input regions.}

\item{seed}{A number used to initialize the random character generator in R, ensuring consistent results.}
}
\description{
The input factors are shuffled and associated with elements. 
The output data.frame contains information about the random shuffle factors and their nearest elements with the distance.
}
