% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{normalDistPeakCompile}
\alias{normalDistPeakCompile}
\title{Compare the observe compile information with expect compile information by normal distribution}
\usage{
normalDistPeakCompile(
  observe,
  expect.data = expect.data,
  parallel = 1,
  parallel.type = "mclapply"
)
}
\arguments{
\item{observe}{A numeric vector of observe data}

\item{expect.data}{A list of numeric vector of expect data}

\item{parallel}{The number of cores to use}

\item{parallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Use mclapply to run in parallel\cr
\cr
"bplapply" - Use BiocParallel to run in parallel}
}
\description{
Compare the observe compile information with expect compile information by normal distribution
}
