% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{NormDistribute}
\alias{NormDistribute}
\title{Create normal distribution table with input numbers}
\usage{
NormDistribute(data)
}
\arguments{
\item{data}{Input numbers}
}
\description{
Create normal distribution table with input numbers
}
