% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{ObsExpBarPlot}
\alias{ObsExpBarPlot}
\title{Plot the observed and expected number in each condition}
\usage{
ObsExpBarPlot(data, width = 0.75, observe.col = "red", expect.col = "grey")
}
\arguments{
\item{data}{A data frame with the following columns:
\itemize{
  \item{condition}{The condition names.}
  \item{observe}{The observed number in each condition.}
  \item{expect}{The expected number in each condition.}
  \item{z_score}{The z-score of the observed number.}
  \item{log2FC}{The log2 fold change of the observed number.}
  \item{upper.p}{The upper p-value of the observed number.}
  \item{lower.p}{The lower p-value of the observed number.}
}}

\item{width}{The width of the bar.}

\item{observe.col}{The color of the observed bar.}

\item{expect.col}{The color of the expected bar.}
}
\value{
A ggplot object.
}
\description{
Plot the observed and expected number in each condition
}
