% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpBinomCompile}
\alias{ObsExpBinomCompile}
\title{Compile the binomial test results.}
\usage{
ObsExpBinomCompile(data)
}
\arguments{
\item{data}{The data should be a list with observe, expect and log2FC, export by ObsExpCompile or ObsExpRDS and contsins the BinomTable.}
}
\description{
Compile the binomial test results.
}
