% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchievePeakStat.R
\name{ObsExpBinomial}
\alias{ObsExpBinomial}
\title{Perform binomial test for the up and down targets.}
\usage{
ObsExpBinomial(data, n = 10, parallel = FALSE)
}
\arguments{
\item{data}{The data should be a list with observe, expect and log2FC, export by ObsExpCompile or ObsExpRDS and contsins the BinomTable.}

\item{n}{The number of bins to perform the binomial test.}

\item{parallel}{If TRUE, use BiocParallel to perform the binomial test.}
}
\description{
Perform binomial test for the up and down targets.
}
