% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpBinomTable}
\alias{ObsExpBinomTable}
\title{Get up or down info from the info of observe and expect to perform binomial test.}
\usage{
ObsExpBinomTable(data)
}
\arguments{
\item{data}{The data should be a list with observe, expect and log2FC, export by ObsExpCompile or ObsExpRDS.}
}
\description{
Get up or down info from the info of observe and expect to perform binomial test.
}
