% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpCompare}
\alias{ObsExpCompare}
\title{TEMP}
\usage{
ObsExpCompare(
  dir = NULL,
  observe = NULL,
  expect.dir = NULL,
  parallel = FALSE,
  shuffle.n = 10000,
  shuffle.prefix = "shuffle",
  observe.prefix = "observe",
  file.ext = ".txt.gz",
  intersect = TRUE,
  condition = c("0-3000", "3000-10000", "10000-20000", "20000-30000", "40000-50000"),
  stat.type = "both",
  tail = "both",
  plotINFO.save = FALSE,
  log.p = TRUE
)
}
\description{
TEMP
}
