% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpCompareByBin}
\alias{ObsExpCompareByBin}
\title{TEMP}
\usage{
ObsExpCompareByBin(
  dir = NULL,
  observe = NULL,
  expect.dir = NULL,
  parallel = FALSE,
  shuffle.n = 10000,
  shuffle.prefix = "shuffle",
  observe.prefix = "observe",
  file.ext = ".txt.gz",
  intersect = TRUE,
  bin = 1000,
  min = 0,
  max = 1e+06,
  count.type = "continue",
  stat.type = "both",
  tail = "both",
  plotINFO.save = FALSE,
  log.p = TRUE
)
}
\description{
TEMP
}
