% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakStat.R
\name{ObsExpCompile}
\alias{ObsExpCompile}
\title{Combine the observe and expect info to do the statistic.}
\usage{
ObsExpCompile(expect, observe = observe)
}
\arguments{
\item{expect}{Expect info.}

\item{observe}{Observe info.}
}
\description{
Combine the observe and expect info to do the statistic.
}
