% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakStat_archive.R
\name{ObsExpCompileList}
\alias{ObsExpCompileList}
\title{Combine the observe and multiple expect info to do the statistic.}
\usage{
ObsExpCompileList(
  expect,
  observe = observe,
  parallel = FALSE,
  expect.name = NULL
)
}
\arguments{
\item{expect}{Expect info.}

\item{observe}{Observe info.}

\item{parallel}{If TRUE, will apply bapply to run the process.}

\item{expect.name}{list of name for expect results. If NULL, will assign as "Shuffle_1", "Shuffle_2"...}
}
\description{
Combine the observe and multiple expect info to do the statistic.
}
