% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakStat_archive.R
\name{ObsExpCompileSTAT}
\alias{ObsExpCompileSTAT}
\title{Perform the statistic for the observe and expect info by exact testing based on edgeR.}
\usage{
ObsExpCompileSTAT(data, type = NULL, parallel = FALSE)
}
\arguments{
\item{data}{The data should be a list with observe, expect and log2FC, export by ObsExpCompile or ObsExpRDS.}

\item{type}{The type of statistic. Could be "binomial", "exact_2tail", "exact_sample_compare", "exact_deviance", "exact_binomial".}

\item{parallel}{If TRUE, will apply bapply to run the process.}
}
\description{
Perform the statistic for the observe and expect info by exact testing based on edgeR.
}
