% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{ObsExpCurvePlot}
\alias{ObsExpCurvePlot}
\title{Plot the observed and expected number in each condition}
\usage{
ObsExpCurvePlot(
  dataList,
  color = NULL,
  xlab = "LogFC (expected/observe)",
  ylab = "Distance of feature to nearest genomic regions",
  hline = 0,
  hline.type = "solid",
  hline.color = "black",
  line.width = 1
)
}
\arguments{
\item{dataList}{A list of data frames with the following columns:
\itemize{
  \item{condition: }{The condition names.}
  \item{observe: }{The observed number in each condition.}
  \item{expect: }{The expected number in each condition.}
  \item{z_score: }{The z-score of the observed number.}
  \item{log2FC: }{The log2 fold change of the observed number.}
  \item{upper.p: }{The upper p-value of the observed number.}
  \item{lower.p: }{The lower p-value of the observed number.}
}}

\item{color}{A vector of colors for each condition.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{hline}{The y-axis value of the horizontal line.}

\item{hline.type}{The type of the horizontal line.}

\item{hline.color}{The color of the horizontal line.}

\item{line.width}{The width of the line.}
}
\value{
A ggplot object.
}
\description{
Plot the observed and expected number in each condition
}
