% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakStat_archive.R
\name{ObsExpIntersectMerge}
\alias{ObsExpIntersectMerge}
\title{Compile the statistic result with intersect numbers from ObsExpCompare.}
\usage{
ObsExpIntersectMerge(data)
}
\arguments{
\item{data}{Result from ObsExpCompare.}
}
\description{
Compile the statistic result with intersect numbers from ObsExpCompare.
}
