% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{ObsExpObj}
\alias{ObsExpObj}
\title{Create the object contained the information of observe and expect.}
\usage{
ObsExpObj(
  factor,
  element = element,
  strand = FALSE,
  tag = FALSE,
  outloc = NULL,
  genome = genome,
  incl = NULL,
  excl = NULL,
  random.n = 10000,
  intersect = TRUE,
  condition = c("0-3000", "3000-10000", "10000-20000", "20000-30000", "40000-50000"),
  parrallel = 1,
  parrallel.type = "mclapply"
)
}
\arguments{
\item{factor}{input the factor data frame or the path to bed file.}

\item{element}{input the element data frame or the path to bed file.}

\item{strand}{if assign as TRUE means input ele,ent contain the strand information at column 6, and will consider the strand information in the analysis.}

\item{tag}{if assign as TRUE means output the tag information in the analysis.}

\item{outloc}{The location of output file.}

\item{genome}{The genome information, could be the path to genome file or data frame contained the genome information.}

\item{incl}{The interval information contained included regions.}

\item{excl}{The interval information contained excluded regions.}

\item{random.n}{Times of shuffle.}

\item{intersect}{If assign as TRUE, result will contained intersect number.}

\item{condition}{Range of distance to nearest factor. Two number saperate by "-".}

\item{parrallel}{If assign number > 1, the function will run in parallel}

\item{parrallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Use mclapply to run in parallel\cr
\cr
"bplapply" - Use BiocParallel to run in parallel}
}
\description{
Create the object contained the information of observe and expect.
}
