% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpSTAT}
\alias{ObsExpSTAT}
\title{Calculate the p-value of observed number of peaks in the expect distribution}
\usage{
ObsExpSTAT(data, log.p = FALSE, parrallel = 1, parrallel.type = "mclapply")
}
\arguments{
\item{data}{The input data contained the information of expect distribution or observe number.}

\item{log.p}{If TRUE, the log of p-value will be returned.}

\item{parrallel}{If assign number > 1, the function will run in parallel}

\item{parrallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Use mclapply to run in parallel\cr
\cr
"bplapply" - Use BiocParallel to run in parallel}
}
\description{
Calculate the p-value of observed number of peaks in the expect distribution
}
