% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{PeaktoPeakIntersect}
\alias{PeaktoPeakIntersect}
\title{Calculate specific peak's (intersect with subject) enrichment factor by compare with random peaks.}
\usage{
PeaktoPeakIntersect(
  subject,
  query = query,
  factor = factor,
  random.num = 10000,
  pval = 0.05,
  parallel = FALSE,
  peak.name = "name"
)
}
\arguments{
\item{subject}{A GRanges object of subject peaks.}

\item{query}{A GRanges object of query peaks.}

\item{factor}{A GRanges object of factor peaks.}

\item{random.num}{Number of random times to get expect result.}

\item{pval}{P-value cutoff.}

\item{parallel}{If TRUE, will use parallel to calculate.}

\item{peak.name}{The column name of peak name in query.}
}
\description{
Calculate specific peak's (intersect with subject) enrichment factor by compare with random peaks.
}
