% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{randomFactor}
\alias{randomFactor}
\title{Randomly select factors from a list.}
\usage{
randomFactor(list, seed = 1, n = NULL)
}
\arguments{
\item{list}{A character vector.}

\item{seed}{Seed number.}

\item{n}{Number of factors to select.}
}
\description{
Randomly select factors from a list.
}
