% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{randomPeak}
\alias{randomPeak}
\title{Radomly select peaks from a GRanges object.}
\usage{
randomPeak(grange, seed = 1, n = NULL, origin.n = NULL)
}
\arguments{
\item{grange}{GRanges object.}

\item{seed}{Seed number.}

\item{n}{Number of peaks to select.}

\item{origin.n}{Number of peaks in grange.}
}
\description{
Radomly select peaks from a GRanges object.
}
