% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse.R
\name{ReconstructExpect}
\alias{ReconstructExpect}
\title{Re parse and perform statistic the shuffle peak sets}
\usage{
ReconstructExpect(
  data,
  from = "within",
  data.type = "old",
  to = c("intersect", "0-3000", "3000-10000", "10000-20000", "20000-30000",
    "40000-50000")
)
}
\arguments{
\item{data}{Input suject}

\item{from}{If assign "within", data type of input data}

\item{data.type}{If assign "old", input data type is old version}

\item{to}{Condition of output data}
}
\description{
Re parse and perform statistic the shuffle peak sets
}
