% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{ScaleBarPlot}
\alias{ScaleBarPlot}
\title{Plot the observed and expected number in each condition}
\usage{
ScaleBarPlot(
  data,
  width = 0.75,
  name = "condition",
  value = NULL,
  sortby = NULL,
  scale = NULL,
  scale_col = c("blue", "white", "red"),
  decreasing = FALSE,
  vertical = FALSE
)
}
\arguments{
\item{data}{A data frame with the following columns:
\itemize{
  \item{condition: }{The condition names.}
  \item{observe: }{The observed number in each condition.}
  \item{expect: }{The expected number in each condition.}
  \item{z_score: }{The z-score of the observed number.}
  \item{log2FC: }{The log2 fold change of the observed number.}
  \item{upper.p: }{The upper p-value of the observed number.}
  \item{lower.p: }{The lower p-value of the observed number.}
}}

\item{width}{The width of the bar.}

\item{name}{The column name of the condition.}

\item{value}{The column name of the value.}

\item{sortby}{The column name to sort the data.}

\item{scale}{The column name to scale the color.}

\item{scale_col}{The color scale.}

\item{decreasing}{Whether to sort the data in decreasing order.}

\item{vertical}{Whether to plot the bar vertically.}
}
\value{
A ggplot object.
}
\description{
Plot the observed and expected number in each condition
}
