% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{ShufFactorElementCorObj}
\alias{ShufFactorElementCorObj}
\title{Shuffle factors across the genome, then associate them to each element with specified distance, 
and compile the information to an object.}
\usage{
ShufFactorElementCorObj(
  element,
  factor = factor,
  dist = 1e+06,
  strand = FALSE,
  parallel = 1,
  genome = genome,
  incl = NULL,
  excl = NULL,
  seed = 1
)
}
\arguments{
\item{element}{Input the element data.frame or the path to bed file.}

\item{factor}{Input the factor data.frame or the path to the bed file.}

\item{dist}{Distance information for associating factors to each element.}

\item{strand}{If set to TRUE, it means that the input element contains strand information in column 6, 
and the analysis will take the strand information into consideration.}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}

\item{genome}{This parameter specifies the data or file path containing the names and sizes of chromosomes or contigs. 
Each name-size pair should be listed on a separate line and delimited by a tab.}

\item{incl}{The interval information to include the input regions.}

\item{excl}{The interval information to exclude the input regions.}

\item{seed}{A number used to initialize the random character generator in R, ensuring consistent results.}
}
\description{
Shuffle factors across the genome, then associate them to each element with specified distance, 
and compile the information to an object.
}
