% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{shuffleCompile}
\alias{shuffleCompile}
\title{Compile the shuffle objects.}
\usage{
shuffleCompile(
  dir,
  shuffle_name = "shuffle_",
  ext_file = ".rds",
  dist = NULL,
  intersect = FALSE,
  include = "all",
  shuffle_times = 100,
  parallel = 1
)
}
\arguments{
\item{dir}{The directory path to the shuffle files.}

\item{shuffle_name}{The prefix of the shuffle files.}

\item{ext_file}{The extension of the shuffle files.}

\item{dist}{Distance information for associating factors to each element.}

\item{intersect}{If set to TRUE, results will include the factor intersect with elements.}

\item{include}{Could be specified one of the: \cr
\cr
"all" - Include all factors associated with elements. \cr
\cr
"upstream" - Include all factors associated with elements at upstream. \cr
\cr
"downstream" -  Include all factors associated with elements at downstream.}

\item{shuffle_times}{The number of shuffle files.}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}
}
\description{
Compile the shuffle objects.
}
