% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromDist.R
\name{shuffleGenerate}
\alias{shuffleGenerate}
\title{Generate shuffled features}
\usage{
shuffleGenerate(factor, genome = genome, incl = NULL, excl = NULL, seed = 1)
}
\arguments{
\item{factor}{Input the factor data.frame or the path to the bed file.}

\item{genome}{This parameter specifies the data or file path containing the names and sizes of chromosomes or contigs. 
Each name-size pair should be listed on a separate line and delimited by a tab.}

\item{incl}{The interval information to include the input regions.}

\item{excl}{The interval information to exclude the input regions.}

\item{seed}{A number used to initialize the random character generator in R, ensuring consistent results.}
}
\description{
Generate shuffled features
}
