% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{TargetFactorSTAT}
\alias{TargetFactorSTAT}
\title{Associate factors to elements in a total list and calculate the significance}
\usage{
TargetFactorSTAT(
  factor,
  total = NULL,
  element = NULL,
  random.num = 10000,
  log.p = FALSE,
  parallel = 1,
  parallel.type = "mclapply"
)
}
\arguments{
\item{factor}{A character vector of factors.}

\item{total}{A character vector of total elements.}

\item{element}{A character vector of elements to be associated.}

\item{random.num}{Number of random times to get expect result.}

\item{log.p}{If TRUE, will log2 the p.value.}

\item{parallel}{If TRUE, will use parallel to calculate.}

\item{parallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Use mclapply to run in parallel\cr
\cr
"bplapply" - Use BiocParallel to run in parallel}
}
\description{
Associate factors to elements in a total list and calculate the significance
}
