% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakStat_archive.R
\name{transformList}
\alias{transformList}
\title{Transform the list of per shuffle with condition to per condition with shuffle}
\usage{
transformList(
  list,
  data = data,
  type = "type",
  number = "number",
  names = TRUE
)
}
\arguments{
\item{list}{Input the characters of conditions (type column in correlation table)}

\item{data}{Correlation table list of expect shuffle peak sets.}

\item{type}{The column name of conditions of range of distance.}

\item{number}{The column name of number of enriched regions.}

\item{names}{If TRUE, will name the output list with type/condition.}
}
\description{
Transform the list of per shuffle with condition to per condition with shuffle
}
