% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{twoFactorElementSTAT}
\alias{twoFactorElementSTAT}
\title{Associate two interval-based factors on the genome within elements, 
compute their correlation, and determine their statistical significance.}
\usage{
twoFactorElementSTAT(
  factorA = NULL,
  factorA.min = 0,
  factorA.max = 0,
  factorB = NULL,
  factorB.min = 0,
  factorB.max = 0,
  element = NULL,
  random.num = 10000,
  log.p = FALSE,
  parallel = 1,
  parallel.type = "mclapply"
)
}
\arguments{
\item{factorA}{A path to a bed file or an intervals information, determine as factorA.}

\item{factorA.min}{Minimum distance of factorA to element. 
If factorA falls within this range of distances, it will be included.}

\item{factorA.max}{Maximum distance of factorA to element. If factorA falls within this range of distances, it will be included.}

\item{factorB}{A path to a bed file or an intervals information, determine as factorB.}

\item{factorB.min}{Minimum distance of factorB to element. 
If factorB falls within this range of distances, it will be included.}

\item{factorB.max}{Maximum distance of factorB to element. 
If factorB falls within this range of distances, it will be included.}

\item{element}{A path to a bed file or an intervals information, determine as element.}

\item{random.num}{Times of randomization.}

\item{log.p}{If set to TRUE, the logarithm of the p-value will be returned.}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}

\item{parallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Apply malapply to perform the run in parallel.\cr
\cr
"bplapply" - Apply bplapply to perfrom the run in parallel.}
}
\description{
Associate two interval-based factors on the genome within elements, 
compute their correlation, and determine their statistical significance.
}
