% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{txdbfromBed}
\alias{txdbfromBed}
\title{Convert bed format data.frame or grange to txdb object}
\usage{
txdbfromBed(data, strand = NULL, name = "name")
}
\arguments{
\item{data}{input data}
}
\description{
Convert bed format data.frame or grange to txdb object
}
