% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{binomialPeakCompile}
\alias{binomialPeakCompile}
\title{Compare the observed compilation information with the expected compilation information using a binomial distribution.}
\usage{
binomialPeakCompile(
  observe,
  expect.data = expect.data,
  p.adjust = TRUE,
  parallel = 1
)
}
\arguments{
\item{observe}{The observe compile information.}

\item{expect.data}{The expect compile information. Typically will be a list.}

\item{p.adjust}{If set to TRUE, the function will adjust the p-value using the FDR method.}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}
}
\description{
Compare the observed compilation information with the expected compilation information using a binomial distribution.
}
