% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{binomialPeakCompile}
\alias{binomialPeakCompile}
\title{Compare the observed compilation information with the expected compilation information using a binomial distribution.}
\usage{
binomialPeakCompile(
  observe,
  expect.data = expect.data,
  p.adjust = TRUE,
  parallel = 1,
  parallel.type = "mclapply"
)
}
\arguments{
\item{observe}{The observe compile information.}

\item{expect.data}{The expect compile information. Typically will be a list.}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}

\item{parallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Apply malapply to perform the run in parallel.\cr
\cr
"bplapply" - Apply bplapply to perfrom the run in parallel.}
}
\description{
Compare the observed compilation information with the expected compilation information using a binomial distribution.
}
