% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse.R
\name{CountCorrelation}
\alias{CountCorrelation}
\title{Input the correlated (annotation) data}
\usage{
CountCorrelation(
  data,
  intersect = TRUE,
  condition = c("0-3000", "3000-10000", "10000-20000", "20000-30000", "40000-50000")
)
}
\arguments{
\item{data}{Input data.frame or compressed/uncompressed data. Data should output from RegionAnnoFromData, contain the information of
name (1st column); $tag_region (2nd column); $tag_type (3rd column); $tag_dist (4th column)}

\item{intersect}{If assign TRUE, output result will contained the number intersect peaks ($tag_dist==0) at first row}

\item{condition}{list of two number saperate with "-", will be used to count the number of factor A within the distance of numbers.}
}
\description{
Input the correlated (annotation) data and count the number of each condition. (e.g. intersect or with specific distances)
}
