% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{CountCorrelationByBin}
\alias{CountCorrelationByBin}
\title{Input the correlated (annotation) data}
\usage{
CountCorrelationByBin(
  data,
  intersect = TRUE,
  bin = 1000,
  min = 0,
  max = 1e+06,
  count.type = "within"
)
}
\arguments{
\item{data}{Input data.frame or compressed/uncompressed file. 
Data should be in the format outputted by RegionAnnoFromData . It contain the information of: \cr
\cr
column 1: Name of the interval for the input factor. \cr
column 2: Name of the interval for the input element. \cr
column 3: Relative position information of factors with respect to the nearest element. \cr
column 4: Distance from the factors to the nearest element.}

\item{intersect}{If set to TRUE, the output result will place the number of intersecting peaks 
(where column 4's "distance" is equal to 0) in the first row.}

\item{bin}{Interval size of each window..}

\item{min}{Minimum number of conditions}

\item{max}{Maximum number of conditions}

\item{type}{Could be specify: \cr
\cr
"continue" - Output a string of intervals with the range of distances between 0 and a series of consecutive numbers. \cr
\cr
"eachBin"   - Output a string of intervals based on specified bin sizes, dividing the range of distances into equal segments.}
}
\description{
Input the correlated (annotation) data and count the number of each condition by bins. (e.g. intersect or with specific distances)
}
