% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{FactorElementCorrelate}
\alias{FactorElementCorrelate}
\title{Input factors and associate to elements. Output data.frame conatined the information of factors to nearest elements with the distance.}
\usage{
FactorElementCorrelate(
  factor,
  element = element,
  strand = FALSE,
  tag = FALSE,
  outloc = NULL
)
}
\arguments{
\item{factor}{input the factor data.frame or the path to bed file.}

\item{element}{input the element data.frame or the path to bed file.}

\item{strand}{if assign as TRUE means input ele,ent contain the strand information at column 6, and will consider the strand information in the analysis.}

\item{tag}{if assign as TRUE means output the tag information in the analysis.}

\item{outloc}{The location of output file.}
}
\description{
Input factors and associate to elements. Output data.frame conatined the information of factors to nearest elements with the distance.
}
