% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchievePeakStat.R
\name{ObsExpCompare}
\alias{ObsExpCompare}
\title{Define the observe and expect annotation info to do the statistic.}
\usage{
ObsExpCompare(
  dir = NULL,
  observe = NULL,
  expect.dir = NULL,
  parallel = FALSE,
  shuffle.n = 10000,
  shuffle.prefix = "shuffle_",
  observe.prefix = "observe",
  file.ext = ".txt.gz",
  intersect = TRUE,
  condition = c("0-3000", "3000-10000", "10000-20000", "20000-30000", "40000-50000"),
  stat.type = "both",
  tail = "both",
  plotINFO.save = FALSE,
  log.p = TRUE
)
}
\arguments{
\item{dir}{Directory contained observe and expect info files.}

\item{observe}{PATH to observe info.}

\item{expect.dir}{Directory contained expect info files.}

\item{parallel}{If TRUE, will apply bapply to run the process.}

\item{shuffle.n}{Times of shuffle.}

\item{shuffle.prefix}{Prefix name of shuffle info.}

\item{observe.prefix}{Prefix name of observe info.}

\item{file.ext}{Extension name of all files.}

\item{intersect}{If TRUE, result will contained intersect number.}

\item{condition}{Range of distance to nearest factor. Two number saperate by "-".}

\item{stat.type}{Statistic method. \cr
\cr
norm: Apply p.norm to calculate the significance \cr
\cr
ecdf : Apply ecdf function to calculate probability of upeer or lower. \cr
\cr
both : Both apply norm and ecdf.}

\item{tail}{Could assign as upper, lower or both.}

\item{plotINFO.save}{If assign as TRUE, will save the essential information to do some plot.}

\item{log.p}{If TRUE, will log2 the p.value (Only could apply on p.norm).}
}
\description{
Define the observe and expect annotation info to do the statistic.
}
