% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDATA.R
\name{ObsExpObjEachBin}
\alias{ObsExpObjEachBin}
\title{Create the object contained the information of observe and expect by each bin.}
\usage{
ObsExpObjEachBin(
  factor,
  element = element,
  strand = FALSE,
  tag = FALSE,
  outloc = NULL,
  genome = genome,
  incl = NULL,
  excl = NULL,
  random.n = 10000,
  intersect = TRUE,
  bin = 1000,
  min = 0,
  max = 1e+06,
  count.type = "within",
  parrallel = FALSE
)
}
\arguments{
\item{factor}{input the factor data frame or the path to bed file.}

\item{element}{input the element data frame or the path to bed file.}

\item{strand}{if assign as TRUE means input ele,ent contain the strand information at column 6, and will consider the strand information in the analysis.}

\item{tag}{if assign as TRUE means output the tag information in the analysis.}

\item{outloc}{The location of output file.}

\item{genome}{The genome information, could be the path to genome file or data frame contained the genome information.}

\item{incl}{The interval information contained included regions.}

\item{excl}{The interval information contained excluded regions.}

\item{random.n}{Times of shuffle.}

\item{intersect}{If assign as TRUE, result will contained intersect number.}

\item{bin}{Interval size of each window.}

\item{min}{Minimum number of conditions}

\item{max}{Maximum number of conditions}

\item{count.type}{Could be specify one of: \cr
\cr
"continue" - All conditions will continues from min+interval*(max/bin(n)-1) to min+interval(max/bin(n))\cr
\cr
"within"   - All conditions will start from 0 to each intervals {0+interval(max/bin(n))}}

\item{parallel}{If assign number > 1, the function will run in parallel}

\item{parrallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Use mclapply to run in parallel\cr
\cr
"bplapply" - Use BiocParallel to run in parallel}
}
\description{
Create the object contained the information of observe and expect by each bin.
}
