% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpRDS}
\alias{ObsExpRDS}
\title{Combine the observe and expect info from RDS files to do the statistic.}
\usage{
ObsExpRDS(
  shuffleRDS,
  observeRDS = observeRDS,
  dist = 1e+06,
  parallel = FALSE,
  expect.name = NULL
)
}
\arguments{
\item{shuffleRDS}{PATH to shuffle info. Could assign as a list of RDS files.}

\item{observeRDS}{PATH to observe info.}

\item{dist}{Distance to calculate the number of peaks.}

\item{parallel}{If TRUE, will apply bapply to run the process.}

\item{expect.name}{list of name for expect results. If NULL, will assign as "Shuffle_1", "Shuffle_2"...}
}
\description{
Combine the observe and expect info from RDS files to do the statistic.
}
