% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpSTAT}
\alias{ObsExpSTAT}
\title{Calculate the p-value by comparing the observed value to an empirical model constructed 
from expected values with each range of distances (condition).}
\usage{
ObsExpSTAT(data, log.p = FALSE, parallel = 1)
}
\arguments{
\item{data}{The input object contained the information of observed and expected number.}

\item{log.p}{If TRUE, the log of p-value will be returned.}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}
}
\description{
Calculate the p-value by comparing the observed value to an empirical model constructed 
from expected values with each range of distances (condition).
}
