% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{ObsExpSTATbyName}
\alias{ObsExpSTATbyName}
\title{Calculate the p-value of observed number of peaks in the expect distribution}
\usage{
ObsExpSTATbyName(data, name = name, log.p = FALSE)
}
\arguments{
\item{data}{The input data contained the information of expect distribution or observe number.}

\item{name}{The name of condition to be calculated.}

\item{log.p}{If TRUE, the log of p-value will be returned.}
}
\description{
Calculate the p-value of observed number of peaks in the expect distribution
}
