% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{randomFactor}
\alias{randomFactor}
\title{Randomly select a character from a character list.}
\usage{
randomFactor(list, seed = 1, n = NULL)
}
\arguments{
\item{list}{The input list contained characters.}

\item{seed}{A number used to initialize the random character generator in R, ensuring consistent results.}

\item{n}{Number of characters to select.}
}
\description{
Randomly select a character from a character list.
}
