% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompilePeak.R
\name{ShufFactorElementCorObj}
\alias{ShufFactorElementCorObj}
\title{Compile the associated factors to each element with distance information by shuffle factors}
\usage{
ShufFactorElementCorObj(
  element,
  factor = factor,
  dist = 1e+06,
  strand = FALSE,
  parallel = 1,
  parallel.type = "mclapply",
  genome = genome,
  incl = NULL,
  excl = NULL,
  seed = 1
)
}
\arguments{
\item{element}{The element file with bed format}

\item{factor}{The factor file with bed format}

\item{dist}{The distance to define the associated factors}

\item{strand}{The strand information of element file}

\item{parallel}{The number of cores to run the function}

\item{parallel.type}{The type of parallel}

\item{genome}{The genome information}

\item{incl}{The included chromosomes}

\item{excl}{The excluded chromosomes}

\item{seed}{The seed number}
}
\description{
Compile the associated factors to each element with distance information by shuffle factors
}
