% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{TargetFactorSTAT}
\alias{TargetFactorSTAT}
\title{Associate factors to elements in a total list and calculate the significance}
\usage{
TargetFactorSTAT(
  factor,
  total = NULL,
  element = NULL,
  random.num = 10000,
  log.p = FALSE
)
}
\arguments{
\item{factor}{A character vector of factors.}

\item{total}{A character vector of total elements.}

\item{element}{A character vector of elements to be associated.}

\item{random.num}{Number of random times to get expect result.}

\item{log.p}{If TRUE, will log2 the p.value.}
}
\description{
Associate factors to elements in a total list and calculate the significance
}
