% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{TargetFactorSTAT}
\alias{TargetFactorSTAT}
\title{Determine significance by associating factors with elements in the complete list.}
\usage{
TargetFactorSTAT(
  factor,
  total = NULL,
  element = NULL,
  random.num = 10000,
  log.p = FALSE,
  parallel = 1,
  parallel.type = "mclapply"
)
}
\arguments{
\item{factor}{Characters determined as factors.}

\item{total}{The complete list is utilized to extract randomized factors for calculating significance.}

\item{element}{Characters determined as element.}

\item{random.num}{Times of randomization.}

\item{log.p}{If set to TRUE, the logarithm of the p-value will be returned.}

\item{parallel}{If a number greater than 1 is assigned, the function will run in parallel.}

\item{parallel.type}{Could be specify one of: \cr
\cr
"mclapply" - Apply malapply to perform the run in parallel.\cr
\cr
"bplapply" - Apply bplapply to perfrom the run in parallel.}
}
\description{
Determine significance by associating factors with elements in the complete list.
}
