% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{twoFactorElementSTAT}
\alias{twoFactorElementSTAT}
\title{Associate two factors in elements and calculate the significance}
\usage{
twoFactorElementSTAT(
  factorA = NULL,
  factorA.min = 0,
  factorA.max = 0,
  factorB = NULL,
  factorB.min = 0,
  factorB.max = 0,
  element = NULL,
  random.num = 10000,
  log.p = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{factorA}{A bed information of factorA.}

\item{factorA.min}{Minimum distance of element to factorA.}

\item{factorA.max}{Maximum distance of element to factorA}

\item{factorB}{A bed information of factorB.}

\item{factorB.min}{Minimum distance of element to factorB.}

\item{factorB.max}{Maximum distance of element to factorB.}

\item{element}{A bed information of elements.}

\item{random.num}{Number of random times to get expect result.}

\item{log.p}{If TRUE, will log2 the p.value.}

\item{parrallel}{If TRUE, will use parallel to calculate.}
}
\description{
Associate two factors in elements and calculate the significance
}
