% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchieveFunction.R
\name{txdbfromBed}
\alias{txdbfromBed}
\title{Convert bed format data.frame or grange to txdb object}
\usage{
txdbfromBed(data, name = "name", strand = NULL)
}
\arguments{
\item{data}{input data}

\item{name}{assign the fourth column name and keep it exsist in grange}

\item{strand}{if strand assign as TRUE means input data contain the strand information at column 6, if assign NULL strand will be *, otherwise could be "+" or "-"}
}
\description{
Convert bed format data.frame or grange to txdb object
}
